ATA:ahci-platform: remove disable/enable sata clock c2k
On Comcerto2k it is possible to disable to SATA clock to save some
energy. But this may causs problems with certain SATA drives, which are
directly 'one'. So SATA bus get out of sync of device/jost must di a
resync, this can lead to problems with underlying layers aka RAID.

Disable the whole clocking in suspend/resume code.
Codepaths in driver init/exit are untouched for bring the PLL to live
while start the host.

Maybe some delay at interface up is also needed.

Signed-off-by: Hans Ulli Kroll <ulli.kroll@googlemail.com>